fixed Popup and ContextMenu having fixed children
